backend/requirements.txt
fastapi
uvicorn
pdfplumber
pandas
python-multipart
pdfminer.six
numpy-financial
pyxirr
asyncpg
psycopg2-binary
google-auth
requests
